refactor: Fixed type check, moved types to a separate file
