+++ package.json
@@ -42,7 +42,7 @@
     "jsdom": "^22.1.0",
     "prettier": "^2.8.7",
     "semver": "^7.6.0",
-    "typescript": "^4.9.5",
+    "typescript": "<5.1.0",
     "vite": "^4.2.0",
     "vite-plugin-glsl": "^1.1.2",
     "vite-plugin-svgr": "^4.0.0",
